+++ MemfaultCloud/Classes/MFLTChunkSender.h
@@ -18,7 +18,8 @@ NS_ASSUME_NONNULL_BEGIN
                           chunkQueue:(id<MemfaultChunkQueue>)chunkQueue
                        dispatchQueue:(dispatch_queue_t)dispatchQueue
                                  api:(MemfaultApi *)api
-                               backoff:(MFLTBackoff *)backoff;
+                               backoff:(MFLTBackoff *)backoff
+              maxConsecutiveErrorCount:(NSInteger)maxConsecutiveErrorCount;
 
 @end